### 2022-07-05 #4547 @editorialbot
[PRE REVIEW]: ThermoFun: A C++/Python library for computing standard thermodynamic properties of substances and reactions across wide ranges of temperatures and pressures
**Submitting author:** @gdmiron (<a href="http://orcid.org/0000-0002-6923-7300">George Miron</a>)
**Repository:** https://github.com/thermohub/thermofun
**Branch with paper.md** (empty if default branch): paper
**Version:** v0.3.9
