### 2022-07-05 #4547 @danielskatz
👋 @gdmiron - it looks like you need to make some changes to your markdown paper where there are links -
`ThermoFun [thermohub.org/thermofun](https://thermohub.org/thermofun) is an open source library` should probably either be `[ThermoFun](https://thermohub.org/thermofun) is an open source library` or `ThermoFun ([thermohub.org/thermofun](https://thermohub.org/thermofun)) is an open source library`  Other links have this same issue.
