### 2022-07-05 #4547 @danielskatz
Also, the paper seems rather long - are there parts that can be removed in favor of pointing to documentation or other papers?
