### 2022-07-06 #4547 @gdmiron
Dear @danielskatz thank you for your comments. I will do the suggested fixes to the links. Concerning the length of the article, ThermoFun is a rather extensive code concerning thermodynamic models and data (database) for substances and reactions. We (the authors) tried to have a summary paper covering the very main features described in a simple and hopefully clear manner (for non experts) without going into any details or technical aspects or more complex use cases. This makes it difficult for us to remove sections of the paper, we are of course open to suggestions.
