### 2022-07-06 a13a289
small fixes based on JOSS editor comments
