+++ paper/paper.md
@@ -30,7 +30,7 @@ bibliography: paper.bib
 
 # Summary
 
-ThermoFun [thermohub.org/thermofun](https://thermohub.org/thermofun) is an open source library that greatly facilitates the computation and retrieval of standard thermodynamic properties of chemical substances, compounds and reactions among them at temperature and pressure ranges of interest. This is done using many thermodynamic models with input standard properties and parameters from various thermodynamic datasets collected and curated in ThermoHub database [thermohub.org/thermohub](https://thermohub.org/thermohub). The library with its API is developed in C++ for performance reasons. A Python API is maintained for broader and more convenient usage. Both single and batch calculations are possible, as well as coupling ThermoFun with other C++ or Python modelling codes that require standard thermodynamic data as input, e.g., for chemical equilibria calculations, plotting phase diagram plotters, or model parameter fitting. Results can be written to file in tabular or other formats, or passed in memory via the C++ or Python API within a coupled code. For its calculations, ThermoFun can use different thermodynamic datasets that are read from local files (extendable) or are retrieved via remote access to ThermoHub online database. The library contains a comprehensive collection of models and equations of state for solid, aqueous, surface, gaseous, and melt substances, and their reactions, arranged in a modular and extendable way permitting new methods and equations of state to be implemented incrementally. Examples for using ThermoFun are available in its open source code repository [github.com/thermohub/thermofun](https://github.com/thermohub/thermofun). This ensures a broad availability of ThermoFun and of the thermodynamic datasets from ThermoHub in Earth and material science and engineering.
+[ThermoFun](https://thermohub.org/thermofun) is an open source library that greatly facilitates the computation and retrieval of standard thermodynamic properties of chemical substances, compounds and reactions among them at temperature and pressure ranges of interest. This is done using many thermodynamic models with input standard properties and parameters from various thermodynamic datasets collected and curated in [ThermoHub](https://thermohub.org/thermohub) database. The library with its API is developed in C++ for performance reasons. A Python API is maintained for broader and more convenient usage. Both single and batch calculations are possible, as well as coupling ThermoFun with other C++ or Python modelling codes that require standard thermodynamic data as input, e.g., for chemical equilibria calculations, plotting phase diagram plotters, or model parameter fitting. Results can be written to file in tabular or other formats, or passed in memory via the C++ or Python API within a coupled code. For its calculations, ThermoFun can use different thermodynamic datasets that are read from local files (extendable) or are retrieved via remote access to ThermoHub online database. The library contains a comprehensive collection of models and equations of state for solid, aqueous, surface, gaseous, and melt substances, and their reactions, arranged in a modular and extendable way permitting new methods and equations of state to be implemented incrementally. Examples for using ThermoFun are available in its open source [code repository](https://github.com/thermohub/thermofun). This ensures a broad availability of ThermoFun and of the thermodynamic datasets from ThermoHub in Earth and material science and engineering.
 
 # Benefits of using ThermoFun
 
@@ -51,19 +51,19 @@ The ThermoFun library as a service to the community that facilitates the computa
 ![Example of standard thermodynamic properties of substances and reactions calculated using ThermoFun as function of temperature along the saturated water vapor pressure curve (input parameters from `aq17` thermodynamic dataset [@Miron2016;@Miron2017]).\label{fig:properties}](figure1.png)
 
[…]
