### 2022-07-06 86db0fb
GEMS-210 Added function for set up logging level and log to file
