+++ ThermoFun/Common/Exception.cpp
@@ -4,12 +4,45 @@
 #include <algorithm>
 #include <sstream>
 #include <spdlog/sinks/stdout_color_sinks.h>
+#include <spdlog/sinks/rotating_file_sink.h>
 #include "ThermoProperties.h"
+
+#define LOG_PATTERN "[%n] [%^%l%$] %v"
 
 namespace ThermoFun {
 
 // Thread-safe logger to stdout with colors
 std::shared_ptr<spdlog::logger> thfun_logger = spdlog::stdout_color_mt("thermofun");
+
+void update_loggers( bool use_cout, const std::string& logfile_name, size_t log_level)
+{
+    auto thermofun_logger = spdlog::get("thermofun");
+    auto chemicalfun_logger = spdlog::get("chemicalfun");
+
+    // change level
+    spdlog::level::level_enum log_lev = spdlog::level::info;
+    if( log_level<7 ) {
+        log_lev = static_cast<spdlog::level::level_enum>(log_level);
+    }
+    thermofun_logger->set_level(log_lev);
+    chemicalfun_logger->set_level(log_lev);
+
+    //change sinks
+    thermofun_logger->sinks().clear();
+    chemicalfun_logger->sinks().clear();
+    if(use_cout) {
+        auto console_output = std::make_shared<spdlog::sinks::stdout_color_sink_mt>();
+        console_output->set_pattern(LOG_PATTERN);
+        thermofun_logger->sinks().push_back(console_output);
+        chemicalfun_logger->sinks().push_back(console_output);
+    }
+    if(!logfile_name.empty()) {
+        auto file_output = std::make_shared<spdlog::sinks::rotating_file_sink_mt>(logfile_name, 1048576, 3);
+        thermofun_logger->sinks().push_back(file_output);
+        chemicalfun_logger->sinks().push_back(file_output);
+    }
+}
+
 
 namespace internal {
 /// Creates the location string from the file name and line number.
