### 2022-07-06 #4547 @gdmiron
Possible reviewers from JOSS list: Richard Gowers, David Hagan, Katy Barnhart, ...
