### 2022-07-06 #4547 @danielskatz
@gdmiron - you're looking at the editor list - we don't generally also want editors to be reviewers, as they are busy editing.  The reviewer list is [here](https://bit.ly/joss-reviewers) (please start at the bottom)
