### 2022-07-06 c0eba1d
GEMS-210 Added function for set up logging level and log to file for Python API
