+++ ThermoFun/GlobalVariables.h
@@ -19,6 +19,8 @@
 
 namespace ThermoFun {
 
+void update_loggers( bool use_cout, const std::string& logfile_name, size_t log_level);
+
 typedef std::vector<double> vd;
 typedef std::vector<std::vector<double> > vvd;
 
