### 2022-07-13 #4547 @lucydot
@gdmiron I can see you fixed the links - excellent.

I agree with Daniel that the paper is rather long and could be trimmed down. See the paper as a short advert for the package itself. Scanning through, I'd say that the Statement of Need section is particularly long - usually here we don't go into the scientific details, but outline the key purpose (usually in a single paragraph - an "elevator pitch") and how this software fills a particular niche, with reference to what work/software already exists . A Basic Example is also not needed - you could provide a link to this instead. For more info on what the paper should contain see here: https://joss.readthedocs.io/en/latest/review_criteria.html.
