### 2022-07-14 #4547 @gdmiron
@lucydot I see your point and that the journal insists on "short" papers. I am really happy that there is an opportunity to publish code without the need of elaborate application/examples. I just feel that that now we are going to the other "extreme" of having just an abstract paper / extended abstract. In the end one can have everything on a documentation site and leave just the title to be published. The Idea is to give the reader a short overview of what is used for, features, and one example without him having to click away. Even for a non specialist to have an idea of what are these properties, why they are important, and why we have this library. I really feel that the value of the paper would be really decreased and just be a doi placeholder. Maybe I am not used with the "software meta paper" format.
