### 2022-07-14 #4547 @lucydot
@gdmiron we can also see what the reviewers think, and come back to this point. If you have not already, I suggest reading a few of the JOSS papers already published - to see how they balance this trade-off.

I am on leave from tomorrow (15th Jul) for one week. I will continue to look for a second reviewer, however please note that my responses may be more delayed than usual.
