### 2022-07-15 003dfd0
adjustments to logger
