### 2022-07-15 de3a95d
Added logger (#33)
* pybind11 msvc compiler bug https://github.com/microsoft/onnxruntime/issues/9735

* changes to use of logger

* GEMS-210 Added function for set up logging level and log to file

* GEMS-210 Added function for set up logging level and log to file for Python API

* updated pipelines CI to macos12

Co-authored-by: svetad <sv.dmytr@gmail.com>
Co-authored-by: G. Dan Miron <dan.miron@psi.com>
