+++ ThermoFun/Substances/Solute/SoluteHKFgems.cpp
@@ -175,14 +175,8 @@ auto gShok2(Reaktoro_::Temperature TC, Reaktoro_::Pressure Pbar, const Propertie
 
     if (D.val >= 1.4)
     {
-        std::cout << "water density higher than 1.4 g*cm-3, Dw = "
-            << ps.density / 1000 << "g*cm-3." << std::endl << __FILE__ << std::endl << __LINE__;
-//        Exception exception;
-//        exception.error << "Error in gShock2";
-//        exception.reason << "water density higher than 1.4 g*cm-3, Dw = "
-//            << D.val << "g*cm-3.";
-//        exception.line = __LINE__;
-//        RaiseError(exception);
+        thfun_logger->warn(" {} {}: water density higher than 1.4 g*cm-3, Dw = {} g*cm-3. Outside the applicability limits of the HKF model.",
+                           __FILE__, __LINE__, static_cast<double>(ps.density/1000) );
     }
 
     // Check if the point (T,P) is inside region III or the shaded region in Fig. 6 of
