+++ ThermoFun/Substances/Gases/s_solmod2_.cpp
@@ -27,10 +27,8 @@
 
 #include <cmath>
 #include <cstdio>
-#include <iostream>
-#include <iomanip>
 #include <fstream>
-using namespace std;
+
 #include "s_solmod_.h"
 #include "verror.h"
 
