+++ azure-pipelines.yml
@@ -19,7 +19,7 @@ jobs:
 
 - job: macOS
   pool:
-    vmImage: 'macOS-10.15'
+    vmImage: 'macos-12'
 
   steps:
   - task: Bash@3

[…]
