+++ CMakeLists.txt
@@ -53,7 +53,7 @@ if(NOT CMAKE_BUILD_TYPE)
 endif()
 
 # Set the C++ standard
-set(CMAKE_CXX_STANDARD 14)
+set(CMAKE_CXX_STANDARD 17)
 set(CMAKE_CXX_STANDARD_REQUIRED ON)
 
 # Set the list of compiler flags for MSVC compiler
