### 2022-07-16 479584c
link to spdlog
