+++ environment.devenv.yml
@@ -11,7 +11,6 @@ dependencies:
   - ccache  # [unix]
   - pybind11
   - nlohmann_json
-  - fmt>=8.1.1
   - chemicalfun>=0.1.3
   - python={{ python_version }}
   - pytest
