+++ ThermoFun/CMakeLists.txt
@@ -32,6 +32,11 @@ if(ChemicalFun_FOUND)
     target_link_libraries(ThermoFun PUBLIC ChemicalFun::ChemicalFun)
 endif()
 
+if(ChemicalFun_FOUND)
+    target_link_libraries(ThermoFun PRIVATE spdlog::spdlog)
+endif()
+
+
 if (CMAKE_CXX_COMPILER_ID MATCHES "MSVC")
     #link_directories("/usr/local/lib")
 else() 
