+++ cmake/modules/ThermoFunFindDeps.cmake
@@ -19,3 +19,10 @@ if(NOT ChemicalFun_FOUND)
 else()
   message(STATUS "Found ChemicalFun v${ChemicalFun_VERSION}")
 endif()
+
+find_package(spdlog REQUIRED)
+if(NOT spdlog_FOUND)
+  message(FATAL_ERROR "spdlog library not found")
+else()
+  message(STATUS "Found spdlog v${spdlog_VERSION}")
+endif()
