### 2022-07-16 9597fa1
fix versions for spdlog and fmt libs$
