+++ cmake/modules/ThermoFunFindDeps.cmake
@@ -22,7 +22,14 @@ endif()
 
 find_package(spdlog REQUIRED)
 if(NOT spdlog_FOUND)
-  message(FATAL_ERROR "spdlog library not found")
+  message(FATAL_ERROR "spdlog not found")
 else()
   message(STATUS "Found spdlog v${spdlog_VERSION}")
+endif()
+
+find_package(fmt REQUIRED)
+if(NOT fmt_FOUND)
+  message(FATAL_ERROR "fmt not found")
+else()
+  message(STATUS "Found fmt v${fmt_VERSION}")
 endif()
