### 2022-07-16 56c96ef
link to spdlog in win and macos
