+++ ThermoFun/CMakeLists.txt
@@ -40,14 +40,17 @@ endif()
 if (CMAKE_CXX_COMPILER_ID MATCHES "MSVC")
     #link_directories("/usr/local/lib")
     if(spdlog_FOUND)
-    target_link_libraries(ThermoFun PRIVATE spdlog::spdlog)
+        target_link_libraries(ThermoFun PRIVATE spdlog::spdlog)
     endif()
 else() 
     if (CMAKE_CXX_COMPILER_ID MATCHES "Clang")
 # build options for MacOS
     if(spdlog_FOUND)
-    target_link_libraries(ThermoFun PRIVATE spdlog::spdlog)
-endif()
+        target_link_libraries(ThermoFun PRIVATE spdlog::spdlog)
+    endif()
+    if(fmt_FOUND)
+        starget_link_libraries(ThermoFun PRIVATE fmt::fmt)
+    endif()
     else() 
 # build options for Linux
 # Find library and software dependencies
