+++ cmake/modules/ThermoFunFindDeps.cmake
@@ -27,9 +27,9 @@ else()
   message(STATUS "Found spdlog v${spdlog_VERSION}")
 endif()
 
-#find_package(fmt REQUIRED)
-#if(NOT fmt_FOUND)
-#  message(FATAL_ERROR "fmt not found")
-#else()
-#  message(STATUS "Found fmt v${fmt_VERSION}")
-#endif()
+find_package(fmt REQUIRED)
+if(NOT fmt_FOUND)
+  message(FATAL_ERROR "fmt not found")
+else()
+  message(STATUS "Found fmt v${fmt_VERSION}")
+endif()
