+++ ci/pipelines/install/install-osx.sh
@@ -31,3 +31,4 @@ echo "The install failed" >&2
 exit 1
 fi
 conda list
+python -c "import thermofun"
