### 2022-07-16 1134f95
define the CMake variable SPDLOG_FMT_EXTERNAL to use external fmt
