+++ ThermoFun/CMakeLists.txt
@@ -45,11 +45,12 @@ if (CMAKE_CXX_COMPILER_ID MATCHES "MSVC")
 else() 
     if (CMAKE_CXX_COMPILER_ID MATCHES "Clang")
 # build options for MacOS
-    if(spdlog_FOUND)
-        target_link_libraries(ThermoFun PRIVATE spdlog::spdlog)
-    endif()
     if(fmt_FOUND)
-        target_link_libraries(ThermoFun PRIVATE fmt::fmt)
+    target_link_libraries(ThermoFun PRIVATE fmt::fmt)
+    endif()
+    if(spdlog_FOUND)
+        set(SPDLOG_FMT_EXTERNAL ON)
+        target_link_libraries(ThermoFun PRIVATE spdlog::spdlog)
     endif()
     else() 
 # build options for Linux
