### 2022-07-19 #4547 @paleolimbot
This looks very cool and I'm excited that it's been developed! I don't have the bandwidth to give this a proper review at this time (but look forward to R bindings someday!).
