### 2022-07-19 #4547 @gdmiron
@paleolimbot, glad to hear about your interest in the work, is there something similar to pybind11 available for making R bindings?
