### 2022-07-21 f01936f
test with new chfun, link to spdlog for all os
