+++ ThermoFun/CMakeLists.txt
@@ -32,26 +32,14 @@ if(ChemicalFun_FOUND)
     target_link_libraries(ThermoFun PUBLIC ChemicalFun::ChemicalFun)
 endif()
 
-#if(fmt_FOUND)
-#    target_link_libraries(ThermoFun PRIVATE fmt::fmt)
-#endif()
+target_link_libraries(ThermoFun PRIVATE spdlog::spdlog)
 
 
 if (CMAKE_CXX_COMPILER_ID MATCHES "MSVC")
     #link_directories("/usr/local/lib")
-    if(spdlog_FOUND)
-        target_link_libraries(ThermoFun PRIVATE spdlog::spdlog)
-    endif()
 else() 
     if (CMAKE_CXX_COMPILER_ID MATCHES "Clang")
 # build options for MacOS
-    if(fmt_FOUND)
-    target_link_libraries(ThermoFun PRIVATE fmt::fmt)
-    endif()
-    if(spdlog_FOUND)
-        set(SPDLOG_FMT_EXTERNAL ON)
-        target_link_libraries(ThermoFun PRIVATE spdlog::spdlog)
-    endif()
     else() 
 # build options for Linux
 # Find library and software dependencies
