+++ ci/pipelines/install/install-osx.sh
@@ -19,18 +19,16 @@ conda devenv
 source activate thermofun
 mkdir build
 cd build
-python_path=$(which python)
 # Configure step
 cmake \
-    -DPYTHON_EXECUTABLE:FILEPATH=$python_path \
     -DCMAKE_BUILD_TYPE=Release \
     -DCMAKE_INSTALL_LIBDIR=lib \
     ..
 make install
+conda list
+python -c "import thermofun"
 if [ $? -eq 1 ]
 then
 echo "The install failed" >&2
 exit 1
 fi
-conda list
-python -c "import thermofun"
