### 2022-07-21 78d72e8
Fixes for macos build (#38)
* set C++ standard to 17

* link to spdlog

Co-authored-by: svetad <sv.dmytr@gmail.com>
Co-authored-by: G. Dan Miron <dan.miron@psi.com>
