+++ ci/pipelines/install/install-osx.sh
@@ -25,9 +25,10 @@ cmake \
     -DCMAKE_INSTALL_LIBDIR=lib \
     ..
 make install
+conda list
+python -c "import thermofun"
 if [ $? -eq 1 ]
 then
 echo "The install failed" >&2
 exit 1
 fi
-conda list
