### 2022-07-25 #4547 @lucydot
Thanks @alejandrogallo for your response - I'll see if there is someone who is more comfortable with the scientific context - though I may be back in contact!

Thanks @paleolimbot for your reply

@hgandhi2411 @fnattino are you available to review this JOSS submission?
