+++ install-dependencies.sh
@@ -6,6 +6,7 @@ EXTN=so
 #sudo rm -f /usr/local/include/nlohmann/json.hpp
 #sudo rm -rf /usr/local/include/eigen3/Eigen
 #sudo rm -rf /usr/local/include/pybind11
+#sudo rm -rf /usr/local/include/spdlog
 #sudo rm -f /usr/local/lib/libChemicalFun.$EXTN
 
 workfolder=${PWD}
@@ -31,21 +32,19 @@ test -f /usr/local/include/nlohmann/json.hpp || {
 		 rm -rf ~/code
 }
 
+# spdlog 
+# if no spdlog installed in /usr/local/lib/ (/usr/local/include/spdlog)
+test -d /usr/local/include/spdlog || || {
 
-# Eigen3 math library (added for building and installing xGEMS)
-# if not installed in /usr/local/include/eigen3)
-test -d /usr/local/include/eigen3/Eigen || {
-
-        # Downloading and unpacking eigen3 source code into ~/code/eigen-code
-        mkdir -p ~/code && cd ~/code && mkdir eigen-code
-        wget -c https://bitbucket.org/eigen/eigen/get/default.tar.bz2
-        tar xvjf default.tar.bz2 -C eigen-code --strip-components 1
-
-        # Building Eigen3 library
-        cd eigen-code && \
+        # Building thermofun library
+        mkdir -p ~/code && \
+                cd ~/code && \
+                git clone https://github.com/gabime/spdlog -b v1.10.0  && \
+                cd spdlog && \
                 mkdir -p build && \
                 cd build && \
-                cmake .. && \
+                cmake .. \
+                make && \
                 sudo make install
 
         # Removing generated build files
@@ -71,6 +70,29 @@ test -d /usr/local/include/pybind11 || {
         cd ~ && \
                  rm -rf ~/code
 }
+
+# Eigen
+# Eigen3 math library (added for building and installing xGEMS)
+# if not installed in /usr/local/include/eigen3)
+test -d /usr/local/include/eigen3/Eigen || {
+
+        # Building yaml-cpp library
+        mkdir -p ~/code && \
+                cd ~/code && \
+                git clone https://gitlab.com/libeigen/eigen.git -b 3.4.0 && \
+                cd eigen && \
+                mkdir -p build && \
+                cd build && \
+                cmake .. \
+                make && \
+                sudo make install
+
+        # Removing generated build files
+        cd ~ && \
+                 rm -rf ~/code
+}
+
+
 
 # ChemicalFun library
 # if no ChemicalFun installed in /usr/local/lib/ (/usr/local/include/ChemicalFun)
