### 2022-09-27 750fcb3
Add nlohmann_json to ThermoFunFindDeps.
Removed else branch that prints found version of dep. This is already printed by default when found.
