### 2022-09-27 d0afe8f
Add GitHub Actions to build/test ThermoFun (Linux/macOS/Windows) on each push & pull request
