+++ .github/workflows/windows.yml
@@ -1,0 +1,77 @@
+name: windows
+
+on:
+  push:
+    branches:
+      - master
+
+  pull_request:
+
+  schedule:
+    - cron: "0 5 * * 1"  # runs at 05:00 UTC on Mondays
+
+jobs:
+  build:
+    name: ${{ matrix.config }} build on ${{ matrix.os }} (Python ${{ matrix.python-version }})
+    runs-on: windows-2019
+
+    strategy:
+      fail-fast: true
+      max-parallel: 4
+      matrix:
+        os: ["windows-2019"]
+        python-version: [3.9]
+        config: ["Release"]
+
+    steps:
+      - uses: actions/checkout@v2
+
+      - name: Timestamp  # Idea from https://tinyurl.com/47wz7vpv
+        id: timestamp
+        shell: cmake -P {0}
+        run: |
+          string(TIMESTAMP current_date "%Y-%m-%d-%H;%M;%S" UTC)
+          message("::set-output name=now::${current_date}")
+
+      - name: Configuring Cache
+        id: cache
+        uses: actions/cache@v2
+        with:
+          path: |
+            C:\Users\runneradmin\clcache
+            C:\Users\runneradmin\conda_pkgs_dir
+          key:
+            ${{ runner.os }}-${{ matrix.config }}-${{ runner.python-version }}-${{ github.ref }}-clcache-${ { steps.timestamp.outputs.now } }
+          restore-keys: |
+            ${{ runner.os }}-${{ matrix.config }}-${{ runner.python-version }}-${{ github.ref }}-clcache-
+
+      - uses: conda-incubator/setup-miniconda@v2
+        with:
+          auto-activate-base: false
+          allow-softlinks: true
+          activate-environment: ''
+          channels: conda-forge
+          channel-priority: true
+          use-only-tar-bz2: true  # IMPORTANT: This needs to be set true for caching of miniconda packages to work properly!
+
+      - name: Configuring Conda Environment
+        env:
+          PYTHON_VERSION: ${{ matrix.python-version }}
+        run: |
+          conda config --set always_yes yes --set changeps1 no
+          conda config --add channels conda-forge
+          conda install conda-devenv=2
+          conda devenv
+
+      - name: Building, Testing & Installing ThermoFun
+        shell: cmd
+        run: |
+          call conda activate thermofun || goto :eof
+          clcache -s || goto :eof
+          clcache -z || goto :eof
+          mkdir .build && cd .build || goto :eof
+          cmake -GNinja .. -DCMAKE_BUILD_TYPE=${{ matrix.config }} || goto :eof
+          ninja || goto :eof
+          ninja install || goto :eof
+          python -c "import thermofun" || goto :eof
+          clcache -s || goto :eof
