### 2022-09-27 677232c
Add `vs2019_win-64` as deps.
This causes MSVC compiler to be activated automatically when the conda environment `thermofun` is activated in Windows.
