### 2022-09-27 c7e0645
Add env with PYTHONPATH: C:\Miniconda\envs\thermofun\Library\Lib\site-packages to windows github action
