### 2022-10-13 29db2a0
Update instructions and thirdpary dependencies install (#41)
* updated readme for conda install

* updated install thirdparty

* Update install-dependencies.sh
