+++ install-dependencies.sh
@@ -6,6 +6,7 @@ EXTN=so
 #sudo rm -f /usr/local/include/nlohmann/json.hpp
 #sudo rm -rf /usr/local/include/eigen3/Eigen
 #sudo rm -rf /usr/local/include/pybind11
+#sudo rm -rf /usr/local/include/spdlog
 #sudo rm -f /usr/local/lib/libChemicalFun.$EXTN
 
 workfolder=${PWD}
@@ -15,7 +16,7 @@ BUILD_TYPE=Debug
 # nlohmann/json
 test -f /usr/local/include/nlohmann/json.hpp || {
 
-	# Building yaml-cpp library
+	# Building json library
 	mkdir -p ~/code && \
                 cd ~/code && \
 		git clone https://github.com/nlohmann/json.git && \
@@ -31,21 +32,19 @@ test -f /usr/local/include/nlohmann/json.hpp || {
 		 rm -rf ~/code
 }
 
+# spdlog 
+# if no spdlog installed in /usr/local/lib/ (/usr/local/include/spdlog)
+test -d /usr/local/include/spdlog || || {
 
-# Eigen3 math library (added for building and installing xGEMS)
-# if not installed in /usr/local/include/eigen3)
-test -d /usr/local/include/eigen3/Eigen || {
-
-        # Downloading and unpacking eigen3 source code into ~/code/eigen-code
-        mkdir -p ~/code && cd ~/code && mkdir eigen-code
-        wget -c https://bitbucket.org/eigen/eigen/get/default.tar.bz2
-        tar xvjf default.tar.bz2 -C eigen-code --strip-components 1
-
-        # Building Eigen3 library
-        cd eigen-code && \
+        # Building thermofun library
+        mkdir -p ~/code && \
+                cd ~/code && \
+                git clone https://github.com/gabime/spdlog -b v1.10.0  && \
+                cd spdlog && \
                 mkdir -p build && \
                 cd build && \
-                cmake .. && \
+                cmake .. \
+                make && \
                 sudo make install
 
         # Removing generated build files
@@ -56,7 +55,7 @@ test -d /usr/local/include/eigen3/Eigen || {
 #Pybind11
 test -d /usr/local/include/pybind11 || {
 
-        # Building yaml-cpp library
+        # Building pybind11 library
         mkdir -p ~/code && \
                 cd ~/code && \
                 git clone https://github.com/pybind/pybind11.git && \
@@ -72,11 +71,34 @@ test -d /usr/local/include/pybind11 || {
                  rm -rf ~/code
 }
 
+# Eigen
+# Eigen3 math library (added for building and installing xGEMS)
+# if not installed in /usr/local/include/eigen3)
+test -d /usr/local/include/eigen3/Eigen || {
+
+        # Building eigen library
+        mkdir -p ~/code && \
+                cd ~/code && \
+                git clone https://gitlab.com/libeigen/eigen.git -b 3.4.0 && \
+                cd eigen && \
+                mkdir -p build && \
+                cd build && \
+                cmake .. \
+                make && \
+                sudo make install
+
+        # Removing generated build files
+        cd ~ && \
+                 rm -rf ~/code
+}
+
+
+
 # ChemicalFun library
 # if no ChemicalFun installed in /usr/local/lib/ (/usr/local/include/ChemicalFun)
 test -f /usr/local/lib/libChemicalFun.$EXTN || {
 
-        # Building thermofun library
+        # Building chemicalfun library
         mkdir -p ~/code && \
                 cd ~/code && \
                 git clone https://bitbucket.org/gems4/chemicalfun.git -b $BRANCH_TFUN  && \
