### 2022-10-25 b7e79be
Set spdlog as PUBLIC dependency, which should be propagated to codes dependent on ThermoFun lib
This is needed because currently spdlog.h is exposed to end-users. I recommend creating a wrapper class Logger and hide dependency on spdlog if this is intended.
