+++ ThermoFun/CMakeLists.txt
@@ -32,7 +32,7 @@ add_library(ThermoFun::ThermoFun ALIAS ThermoFun)
 # Link ThermoFun against all required dependencies
 target_link_libraries(ThermoFun
     PUBLIC ChemicalFun::ChemicalFun
-    PRIVATE spdlog::spdlog
+    PUBLIC spdlog::spdlog
     PRIVATE nlohmann_json::nlohmann_json
 )
 
