### 2022-10-25 232196d
Fixed shell script to build spdlog with -fPIC
