### 2022-10-25 d11102d
Update README.md (#45)
* Update README.md

* Update install-dependencies.sh

* Using `<spdlog/spdlog.h>` instead of `"spdlog/spdlog.h"`

* Set spdlog as PUBLIC dependency, which should be propagated to codes dependent on ThermoFun lib

This is needed because currently spdlog.h is exposed to end-users. I recommend creating a wrapper class Logger and hide dependency on spdlog if this is intended.

* Update environment.devenv.yml to chemicalfun 0.1.5

Co-authored-by: Allan Leal <allanmulin@gmail.com>
