### 2022-10-31 #4624 @fnattino
Hi @lucydot, apologies for not having answered earlier. I could review the manuscript in a time frame of ~2 weeks  - would that work?
