### 2022-11-21 e4671ab
Adding fmt as deps to check if CI succeeds.
