+++ environment.devenv.yml
@@ -14,6 +14,7 @@ dependencies:
   - pybind11
   - nlohmann_json
   - spdlog=1.10.0
+  - fmt
   - chemicalfun=0.1.5
   - python={{ python_version }}
   - pytest
