### 2022-11-24 ad95567
fixing spdlog link
