+++ environment.devenv.yml
@@ -5,7 +5,7 @@ name: thermofun
 dependencies:
   - setuptools=59.7.0
   - pip
-  - gxx_linux-64  # [linux]
+  - gxx_linux-64=9.4.0  # [linux]
   - cmake>=3.13
   - ninja
   - clcache  # [win]
@@ -13,8 +13,9 @@ dependencies:
   - clangxx_osx-64  # [osx]
   - pybind11
   - nlohmann_json
-  - spdlog=1.10.0
-  - chemicalfun=0.1.5
+ # - spdlog=1.11.0
+ # - fmt=9.1.0
+  - chemicalfun=0.1.6
   - python={{ python_version }}
   - pytest
   - vs2019_win-64  # [win]
