+++ ci/pipelines/install/install-osx.sh
@@ -27,8 +27,10 @@ cmake \
 make install
 conda list
 python -c "import thermofun"
+cd ..
+pytest -ra -vv --color=yes .
 if [ $? -eq 1 ]
 then
 echo "The install failed" >&2
 exit 1
-fi
+
