+++ environment.devenv.yml
@@ -1,11 +1,11 @@
 name: thermofun
 
-{% set python_version = os.environ.get("PY_VER", "3.7") %}
+{% set python_version = os.environ.get("PY_VER", "3.9") %}
 
 dependencies:
   - setuptools=59.7.0
   - pip
-  - gxx_linux-64=9.4.0  # [linux]
+  - gxx_linux-64  # [linux]
   - cmake>=3.13
   - ninja
   - clcache  # [win]
