+++ pytests/test_thermoengine.py
@@ -22,14 +22,14 @@ class TestThermoEngine(unittest.TestCase):
     def test_properties_reaction_from_equation(self):
         assert self.engine.thermoPropertiesReaction(298.15, 1e5, "Cal = Ca+2 + CO3-2").log_equilibrium_constant.val == pytest.approx(-8.48014, 1e-5, 1e-14)
         assert self.engine.thermoPropertiesReaction(298.15, 1e5, "Al+3 + 4 H2O@ + 0Ca+2= 1Al(OH)4- + 4 \n H+").log_equilibrium_constant.val == pytest.approx(-22.3085, 1e-5, 1e-14)
-        assert self.engine.thermoPropertiesReaction(423.15, 0, "Cal = Ca+2 + CO3-2").log_equilibrium_constant.val == pytest.approx(-10.10169, 1e-5, 1e-14)
+        assert self.engine.thermoPropertiesReaction(423.15, 0, "Cal = Ca+2 + CO3-2").log_equilibrium_constant.val == pytest.approx(-10.10141, 1e-5, 1e-14)
         assert self.engine.thermoPropertiesReaction(873.15, 3000e5, "Cal = Ca+2 + CO3-2").reaction_heat_capacity_cp.val == pytest.approx(-718.75763, 1e-5, 1e-14)
         assert self.engine.thermoPropertiesReaction(348.15, 1e5, "Cal = Ca+2 + CO3-2").reaction_gibbs_energy.ddt == pytest.approx(
               -self.engine.thermoPropertiesReaction(348.15, 1e5, "Cal = Ca+2 + CO3-2").reaction_entropy.val, 1e-5, 1e-14)
 
     def test_properties_reaction(self):
         assert self.engine2.thermoPropertiesReaction(298.15, 1e5, "Meionite-Ca").log_equilibrium_constant.val == pytest.approx(80.873916, 1e-5, 1e-14)
-        assert self.engine2.thermoPropertiesReaction(423.15, 0, "Meionite-Ca").log_equilibrium_constant.val == pytest.approx(28.243799, 1e-5, 1e-14)
+        assert self.engine2.thermoPropertiesReaction(423.15, 0, "Meionite-Ca").log_equilibrium_constant.val == pytest.approx(28.23, 1e-5, 1e-14)
         assert self.engine2.thermoPropertiesReaction(298.15, 0, "Meionite-Ca + 25H+ = HCO3- + 6H4SiO4@ + 4Ca+2 + 6Al+3").log_equilibrium_constant.val == pytest.approx(80.873916, 1e-5, 1e-14)
         assert self.engine2.thermoPropertiesReaction(873.15, 5000e5, "Meionite-Ca + 25H+ = HCO3- + 6H4SiO4@ + 4Ca+2 + 6Al+3").reaction_heat_capacity_cp.val == pytest.approx(-1072.969936, 1e-5, 1e-14)
         assert self.engine2.thermoPropertiesReaction(873.15, 5000e5, "Meionite-Ca + 25H+ = HCO3- + 6H4SiO4@ + 4Ca+2 + 6Al+3").reaction_entropy.val == pytest.approx(-1704.130129, 1e-5, 1e-14)
