### 2022-11-26 6d4ec1f
Removed option `use-only-tar-bz2: true` in GitHub workflows.
