+++ .github/workflows/windows.yml
@@ -24,7 +24,7 @@ jobs:
         config: ["Release"]
 
     steps:
-      - uses: actions/checkout@v2
+      - uses: actions/checkout@v3
 
       - name: Timestamp  # Idea from https://tinyurl.com/47wz7vpv
         id: timestamp
@@ -35,9 +35,7 @@ jobs:
 
       - name: Configuring Cache
         id: cache
-        uses: actions/cache@v2
-        env:
-          CACHE_NUMBER: 1  # Increase this value to reset cache if etc/example-environment.yml has not changed
+        uses: actions/cache@v3
         with:
           path: |
             C:\Users\runneradmin\clcache
@@ -54,7 +52,6 @@ jobs:
           activate-environment: ''
           channels: conda-forge
           channel-priority: true
-          use-only-tar-bz2: true  # IMPORTANT: This needs to be set true for caching of miniconda packages to work properly!
 
       - name: Configuring Conda Environment
         env:
