+++ .github/workflows/osx.yml
@@ -23,7 +23,7 @@ jobs:
         config: ["Release"]
 
     steps:
-      - uses: actions/checkout@v2
+      - uses: actions/checkout@v3
 
       - name: Timestamp  # Idea from https://tinyurl.com/47wz7vpv
         id: timestamp
@@ -34,9 +34,7 @@ jobs:
 
       - name: Configuring Cache
         id: cache
-        uses: actions/cache@v2
-        env:
-          CACHE_NUMBER: 1  # Increase this value to reset cache if etc/example-environment.yml has not changed
+        uses: actions/cache@v3
         with:
           path: |
             /Users/runner/Library/Caches/ccache
@@ -52,7 +50,6 @@ jobs:
           python-version: ${{ matrix.python-version }}
           channels: conda-forge
           channel-priority: true
-          use-only-tar-bz2: true  # IMPORTANT: This needs to be set true for caching of miniconda packages to work properly!
 
       - name: Configuring Conda Environment
         shell: bash -l {0}
