### 2022-11-26 fad1a03
Fix for spdlog link and tests (#50)
* fix link to spdlog/fmt through chemicalfun 

* fix for tests at Psat

* fixed tests

* fix macos ci

* Update environment.devenv.yml

* revert

* Increment CACHE_NUMBER to reset cached conda packages

* Removed option `use-only-tar-bz2: true` in GitHub workflows.

Co-authored-by: svetad <sv.dmytr@gmail.com>
Co-authored-by: G. Dan Miron <dan.miron@psi.com>
Co-authored-by: Allan Leal <allanmulin@gmail.com>
