### 2022-11-28 thermohub/thermofun#52 @hgandhi2411
[JOSS] Doc strings for python api
The documentation on thrrmofun.org is good and also has examples. But it would be more accessible if you had doc strings for different methods available. Right now it's hard to tell what methods are available besides those in the example.
