### 2022-11-28 thermohub/thermofun#53 @fnattino
[JOSS] - contributing guidelines
Thanks for providing indications on how to get support (via gitter)  - however I could not find guidelines on how to file issues and submit contributions - maybe you could add a CONTRIBUTING file to the repository (e.g. see [here](https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/creating-a-default-community-health-file)) and link it within the README file?
