### 2022-11-28 thermohub/thermofun#54 @fnattino
[JOSS] - Python snippet in README
When running the second Python snippet from the README file, I get the following error:
`AttributeError: module 'thermohubclient' has no attribute 'setDatabaseConnectionFilePath'`.

If I comment out the first line, I still get the following:
`AttributeError: 'thermohubclient.PyThermoHubClient.DatabaseClient' object has no attribute 'recordsFromThermoDataSet'`.

Minor point: it would also be nice to inform that `thermohubclient` needs to be installed as a separate dependency to run these examples.
