### 2022-11-28 thermohub/thermofun#52 @fnattino
I agree with @hgandhi2411 that API documentation, as part of both docstrings and online documentation, would be very helpful!
