### 2022-11-28 thermohub/thermofun#55 @fnattino
[JOSS] - state of the field 
The manuscript submitted to JOSS reads well and nicely summarises the functionality of Thermofun.
However, no reference to  other software packages with similar scope is made in the draft. A quick search brought up packages like [Thermo](https://github.com/CalebBell/thermo) and [PYroMat](http://pyromat.org/about.html)  - do these or other packages provide similar functionalities as Thermofun?
