### 2022-11-28 #4624 @fnattino
Hi @lucydot, apologies for the delay. Nice work @gdmiron! I have carried out my review and submitted my comments via the following issues:
- https://github.com/thermohub/thermofun/issues/52
- https://github.com/thermohub/thermofun/issues/53
- https://github.com/thermohub/thermofun/issues/54
- https://github.com/thermohub/thermofun/issues/55
