### 2022-12-01 87ce190
updated readme based on JOSS review suggestions
- added table of content
- added instructions for reporting bugs/issues and contributing
