+++ README.md
@@ -5,6 +5,15 @@ Linux, OSX, Windows
 
 A code for calculating the standard state thermodynamic properties of substances and reactions at a given temperature and pressure.
 
+- [Simple C++ API example](#simple-c-api-example)
+- [Try ThermoFun](#try-thermofun-in-your-browser-click-launch-binder)
+- [Python example](#simple-python-api-example)
+- [Install using Conda](#installation-using-conda)
+- [Build and install using cmake](#install-thermofun-using-cmake)
+- [Build and install using cmake and conda](#compile-and-install-thermofun-using-cmake-and-conda)
+- [Reporting bugs](#reporting-bugs)
+- [Contributing](#contributing-with-development)
+
 ## Try ThermoFun in your browser click launch binder
 
 [![Binder](https://mybinder.org/badge_logo.svg)](https://mybinder.org/v2/gh/thermohub/thermofun-jupyter/master?urlpath=lab/)
@@ -89,7 +98,7 @@ int main()
 
 ```
 
-### Simple Python API example
+## Simple Python API example
 
 * Using a json database file
 
@@ -199,7 +208,7 @@ Conda can be installed from [Miniconda](https://conda.io/miniconda.html).
 sudo apt-get install g++ cmake git
 ```
 
-## Download ThermoFun source code using git clone
+* Download ThermoFun source code using git clone
 
 * In a terminal, at the home directory level e.g. ```<user>@ubuntu:~$``` copy-paste and run the following code:
 
@@ -210,11 +219,11 @@ git clone https://github.com/thermohub/thermofun.git && cd thermofun
 
 * In the terminal you should be in ```~/thermofun$```.
 
-## (A) Build and install ThermoFun library (working with json database files)
+### (A) Build and install ThermoFun library (working with json database files)
 
 This option allows the user to build thermofun library that works with a user provided thermodynamic database file in json format and has only one thirdpary library dependency. To build thermofun with access to the thermohub thermodynamic database cloud and local server see bellow. 
 
-### Install Dependencies (if not using Conda environment)
+#### Install Dependencies (if not using Conda environment)
 
 The thermofun library uses nlohmann/json.hpp as thirdparty dependency to parse database files in json format. To install the header only json library in a terminal ```~/thermofun$``` execute the following: 
 
@@ -223,7 +232,7 @@ The thermofun library uses nlohmann/json.hpp as thirdparty dependency to parse d
 sudo ./install-dependencies.sh
 ```
 
-### Compiling the C++ library
+#### Compiling the C++ library
 
 In the terminal ```~/thermofun$```, execute the following commands:
 
@@ -270,11 +279,11 @@ then execute:
 sudo make install 
 ```
 
-## (B) Build and install ThermoFun library (working with access to the local and cloud ThemroHub database)
+### (B) Build and install ThermoFun library (working with access to the local and cloud ThemroHub database)
 
 This option builds thermofun library together with the dbclient, which provides access to the local and cloud thermohub databases, allowing specific a ThermoDataSet to be used or a selection on elements of the thermodynamic data.
 
-### Install ThermoHubClient
+#### Install ThermoHubClient
 
 Clone and install ThermoHubClient library 
 
@@ -296,7 +305,7 @@ For a global installation of the compiled library in your system, execute:
 sudo make install 
 ```
 
-## Install ThermoFun using CMake and Conda 
+## Compile and install ThermoFun using CMake and Conda 
 
 This procedure uses Conda for handling all the dependencies of ThermoFun and builds ThermoFun for Windows, Mac OS X, and Linux. 
 
@@ -332,3 +341,18 @@ conda activate thermofun
[…]
