### 2022-12-01 a4e1779
Fixed code examples in readme
