### 2022-12-21 #4624 @gdmiron
@lucydot I plan to finish by the time you are back from holidays. Fixes to the manuscript text and acces to minimal documentation. Happy holidays!
