### 2022-12-22 4135898
fixes to manuscript text, added table with similar codes
