+++ paper/paper.md
@@ -30,40 +30,40 @@ bibliography: paper.bib
 
 # Summary
 
-[ThermoFun](https://thermohub.org/thermofun) is an open source library that greatly facilitates the computation and retrieval of standard thermodynamic properties of chemical substances, compounds and reactions among them at temperature and pressure ranges of interest. This is done using many thermodynamic models with input standard properties and parameters from various thermodynamic datasets collected and curated in [ThermoHub](https://thermohub.org/thermohub) database. The library with its API is developed in C++ for performance reasons. A Python API is maintained for broader and more convenient usage. Both single and batch calculations are possible, as well as coupling ThermoFun with other C++ or Python modelling codes that require standard thermodynamic data as input, e.g., for chemical equilibria calculations, plotting phase diagram plotters, or model parameter fitting. Results can be written to file in tabular or other formats, or passed in memory via the C++ or Python API within a coupled code. For its calculations, ThermoFun can use different thermodynamic datasets that are read from local files (extendable) or are retrieved via remote access to ThermoHub online database. The library contains a comprehensive collection of models and equations of state for solid, aqueous, surface, gaseous, and melt substances, and their reactions, arranged in a modular and extendable way permitting new methods and equations of state to be implemented incrementally. Examples for using ThermoFun are available in its open source [code repository](https://github.com/thermohub/thermofun). This ensures a broad availability of ThermoFun and of the thermodynamic datasets from ThermoHub in Earth and material science and engineering.
+[ThermoFun](https://thermohub.org/thermofun) is an open source library that simplifies the calculation and retrieval of standard thermodynamic properties of chemical substances, compounds, and reactions among them at a range of temperatures and pressures. The library is developed in C++ for performance, but it also has a Python API for broader and more convenient usage. It employs a variety of thermodynamic models and equations of state for solid, aqueous, surface, gaseous, and molten substances, and their reactions with input properties and parameters from various thermodynamic datasets that are collected and curated in the in [ThermoHub](https://thermohub.org/thermohub) database. The library can be used as a standalone code for searching and tabulating thermodynamic properties or linked to other modeling codes that require standard thermodynamic data as input. It offers the flexibility to use different thermodynamic datasets, including custom datasets and datasets retrieved from the online ThermoHub database, and to choose the most suitable models for different classes of substances necessary in various modeling applications. It can serve as a common source of thermodynamic models for standard properties of substances and reactions that can be easily integrated and combined, significantly improving the modeling capabilities for diverse (geo)chemical systems and over wide ranges of conditions.
 
 # Benefits of using ThermoFun
 
-* It facilitates and simplifies the computation of standard thermodynamic properties of substances and reactions at temperature and pressure of interest.
-* It employs many equations of state and other thermodynamic models needed for calculating standard properties at different conditions; the user does not need to implement and test them separately in own codes or spreadsheets.
-* It uses input standard properties and parameters from various thermodynamic datasets consistently kept in the (online) ThermoHub database, thus removing the burden of collecting all necessary standard thermodynamic data from vast literature or of writing complex scripts for importing these data from different formats.
-* The available thermodynamic models in ThermoFun are implemented and tested based on their original publications; the thermodynamic data found in ThermoHub is curated and documented by experts, allowing the modeler to save time and effort for a more efficient use of geochemical code tools to address scientific problems.
-* It can serve as a common source of thermodynamic models for standard properties of substances and reactions that can be easily integrated and combined, significantly improving the modeling capabilities for diverse (geo)chemical systems and over wide ranges of conditions.
[…]
