+++ paper/paper.bib
@@ -30,7 +30,7 @@
 }
 @article{Miron2020,
    abstract = {Most of the available thermodynamic data concerning radioactive waste disposal are restricted to values of reaction equilibrium constants (logK298°) at 25 °C and 1 bar. Simple estimation methods such as isocoulombic reactions can be used for extrapolating the properties of reactions involving aqueous species and minerals to elevated temperatures. The aim of this study was to validate the applicability of various alternative isocoulombic reactions to estimate logKT° values of aqueous complexation reactions for lanthanides and actinides to elevated temperatures while taking advantage of new additional literature data, and to identify criteria for choosing the “best” reactions. For each chemical species of interest, a systematic approach using dedicated software and database allowed us to identify the isocoulombic reactions and types of extrapolation that yield the best estimates of standard thermodynamic properties at elevated temperatures, when very limited or no experimental data are available. We have tested aqueous complexation reactions for selected lanthanides and actinides of different valences with chloride, fluoride, sulfate, carbonate, nitrate, phosphate and silicate ligands. “Model” complexation reactions, having known temperature trends, were systematically combined with complex formation reactions of interest whose temperature trends are unknown, into many alternative isocoulombic reactions. For each ion, we investigated which of the generated isocoulombic reactions provide the best estimates for logeKT° of the reaction of interest at elevated temperatures in order to compile the guidelines for choosing the optimal ones, then applying these guidelines to “prediction” subsets. In most cases, knowing only logeKT° at 25 °C (for the reaction of interest), it was possible to obtain rather accurate estimates of logeKT° values at elevated temperatures using isocoulombic reactions that exchange ions with similar charge and hydration properties (hydrated ionic radius and structure of the hydration shell) and known logmKT° of model reactions. These ions and their complexes interact with the solvent in comparable ways, so that their similar heat capacity and entropy effects largely cancel out on both sides of an “optimal” isocoulombic reaction.},
-   author = {G. D. Miron and Dmitrii A. Kulik and Tres Thoenen},
+   author = {G.bell D. Miron and Dmitrii A. Kulik and Tres Thoenen},
    doi = {10.1016/j.gca.2020.07.020},
    issn = {00167037},
    journal = {Geochimica et Cosmochimica Acta},
@@ -96,6 +96,149 @@
    volume = {317},
    year = {2017},
 }
+@article{JOHNSON1992,
+title = {SUPCRT92: A software package for calculating the standard molal thermodynamic properties of minerals, gases, aqueous species, and reactions from 1 to 5000 bar and 0 to 1000°C},
+journal = {Computers & Geosciences},
+volume = {18},
+number = {7},
+pages = {899-947},
+year = {1992},
+issn = {0098-3004},
+doi = {https://doi.org/10.1016/0098-3004(92)90029-Q},
+url = {https://www.sciencedirect.com/science/article/pii/009830049290029Q},
+author = {James W. Johnson and Eric H. Oelkers and Harold C. Helgeson},
+keywords = {SUPCRT92, Equations of state, Standard molal thermodynamic properties, Chemical equilibrium, Minerals, Gases, Aqueous species, HO, Thermodynamics, Geochemistry},
[…]
