### 2022-12-22 thermohub/thermofun#55 @gdmiron
Manuscript was updated with the following section

# Codes with similar functionalities

ThermoFun is an open source library that serves as a common source of ever-expanding thermodynamic models for standard properties of substances and reactions. If ThermoFun does not meet your needs, there are other alternatives that offer similar functionality, potentially using different methods, and sometimes embedded in larger applications. The following list is of codes that share some functionality with ThermoFun, but is not exhaustive:

| Name  | Language  |  Reference |  Source |
|-------|-----------|------------|-------|
| Cantera | C++, Python, Matlab | [cantera] | [Open](https://github.com/Cantera/cantera) |
| CHNOSZ | R, Python | [Dick2019] | [Open](https://chnosz.net/) |
| coolprop | C++ | [Bell2014] | [Open](https://github.com/CoolProp/CoolProp)|
| DEWPython | Python | [Chan2021] | [Open](https://github.com/chandr3w/DEWPython) |
| GeoTPD | R-Shiny | [Bastrakov2019] | [Closed](https://geoscienceaustralia.shinyapps.io/GeoTPD/) |
| PyGCC | Python | [AWOLAYO2022]  | [Open](https://bitbucket.org/Tutolo-RTG/pygcc/src/master/) |
| pyromat | Python | [Martin2022] | [Open](https://github.com/chmarti1/PYroMat)|
| Reaktoro | C++, Python | [Leal2017] | [Open](https://github.com/reaktoro) |
| REFPPROP | FORTRAN  | [Huber2022] | [Closed](https://www.nist.gov/srd/refprop) |
| SteamTables | C++, Web | [megawatsoft] | [Closed](https://www.steamtablesonline.com/steam97web.aspx) |
| SUPCRT92 | Fortran  |  [JOHNSON1992] | [Open](https://github.com/justinnhli/supcrt92) |
| SUPCRTBL | Fortran | [ZIMMER2016]  | [Closed](https://models.earth.indiana.edu/supcrtbl.php) |
| Thermo | Python | [bell] | [Open](https://github.com/CalebBell/thermo)|
| ThermoEngine | C++, Python | [thermoengine] | [Open](https://gitlab.com/ENKI-portal/ThermoEngine) |
