### 2022-12-22 2b1eeaf
Add bibentry for `Bell:Deiters:Leal:2022`
