+++ paper/paper.bib
@@ -1,3 +1,16 @@
+@article{Bell:Deiters:Leal:2022,
+	abstract = {This work uses advanced numerical techniques (complex differentiation and automatic differentiation) to efficiently and accurately compute all the required thermodynamic properties of an equation of state without any analytical derivatives─particularly without any handwritten derivatives. It avoids the tedious and error-prone process of symbolic differentiation, thus allowing for more rapid development of new thermodynamic models. The technique presented here was tested with several equations of state (van der Waals, Peng-Robinson, Soave-Redlich-Kwong, PC-SAFT, and cubic-plus-association) and high-accuracy multifluid models. A minimal set of algorithms (critical locus tracing and vapor-liquid equilibrium tracing) were implemented in an extensible and concise open-source C++ library: teqp (for Templated EQuation of state Package). This work demonstrates that highly complicated equations of state can be implemented faster yet with minimal computational overhead and negligible loss in numerical precision compared with the traditional approach that relies on analytical derivatives. We believe that the approach outlined in this work has the potential to establish a new computational standard when implementing computer codes for thermodynamic models.},
+	author = {Bell, Ian H. and Deiters, Ulrich K. and Leal, Allan M.M.},
+	doi = {10.1021/acs.iecr.2c00237},
+	file = {:C:/Users/Allan/Documents/Mendeley Desktop/Bell, Deiters, Leal - 2022 - Implementing an Equation of State without Derivatives teqp.pdf:BINARY},
+	issn = {15205045},
+	journal = {Industrial and Engineering Chemistry Research},
+	number = {17},
+	pages = {6010--6027},
+	title = {{Implementing an Equation of State without Derivatives: teqp}},
+	volume = {61},
+	year = {2022}
+}
 @article{Kulik2013,
    abstract = {Reactive mass transport (RMT) simulation is a powerful numerical tool to advance our understanding of complex geochemical processes and their feedbacks in relevant subsurface systems. Thermodynamic equilibrium defines the baseline for solubility, chemical kinetics, and RMT in general. Efficient RMT simulations can be based on the operator-splitting approach, where the solver of chemical equilibria is called by the mass transport part for each control volume whose composition, temperature, or pressure has changed. Modeling of complex natural systems requires consideration of multiphase-multicomponent geochemical models that include nonideal solutions (aqueous electrolytes, fluids, gases, solid solutions, and melts). Direct Gibbs energy minimization (GEM) methods have numerous advantages for the realistic geochemical modeling of such fluid-rock systems. Substantial improvements and extensions to the revised GEM interior point method algorithm based on Karpov's convex programming approach are described, as implemented in the GEMS3K C/C++ code, which is also the numerical kernel of GEM-Selektor v.3 package (http://gems.web.psi.ch). GEMS3K is presented in the context of the essential criteria of chemical plausibility, robustness of results, mass balance accuracy, numerical stability, speed, and portability to high-performance computing systems. The stand-alone GEMS3K code can treat very complex chemical systems with many nonideal solution phases accurately. It is fast, delivering chemically plausible and accurate results with the same or better mass balance precision as that of conventional speciation codes. GEMS3K is already used in several coupled RMT codes (e.g., OpenGeoSys-GEMS) capable of high-performance computing. © 2012 Springer Science+Business Media B.V.},
    author = {Dmitrii A. Kulik and Thomas Wagner and Svitlana V. Dmytrieva and Georg Kosakowski and Ferdinand F. Hingerl and Konstantin V. Chudnenko and Urs R. Berner},
