### 2022-12-22 9902e68
GEM-Selektor to codes with similar functionalities
Added GEM-Selektor to the list of codes with similar functionalities, fixed style in the paragraph above.
