### 2022-12-22 5cb460d
Added paragraph under Features on automatic differentiation
