+++ paper/paper.md
@@ -65,6 +65,8 @@ The `ThermoBatch` class can be used to run batch calculations for a list of subs
 
 Reaction properties can be calculated for any `reaction` record present in the thermodynamic dataset, with the extrapolation to the given temperature and pressure done using the model code and its input parameters available in the record. Alternatively, the user can provide strings of reaction equations written using symbols of substances available in the thermodynamic dataset (e.g., Calcite $\leftrightharpoons$ Ca$^{2+}$ + CO$_{3}^{2-}$). In this case, the standard properties of reactions are calculated from the properties of reactants, each of which is first computed for the T, P of interest using the models and parameters defined in their respective `substance` records. Another option is to set a reaction with known properties to define the properties of a given substance. In this case, the standard properties of the reaction-defined substance at T, P are retrieved from the standard properties of the reaction calculated at T, P, along with the standard properties of the other reactants (substances) that are involved in the reaction. Thus, one thermodynamic dataset can contain a combination of substances, reaction-dependent substances, and reactions, whose consistent standard properties at T, P of interest are computed recursively.
 
+ThermoFun takes advantage of automatic differentiation for efficient and convenient calculation of derivatives. This feature allows ThermoFun to produce not only the standard thermodynamic properties of species and reactions, but also their temperature and pressure derivatives. This is necessary, for example, when chemical equilibrium problems where temperature and/or pressure are unknown, and Newton method is used to calculate these variables together with the amounts of species in equilibrium, as is done in Reaktoro [@ Leal2016]. Without temperature and pressure derivatives of standard Gibbs energies, the algorithm would not be fast and robust for these specific calculations. We plan to use the `autodiff` C++ library [@Leal:2018] in ThermoFun for even more capable autodiff capabilities, as it has been used in both Reaktoro [@Leal2016] and TEQP [@Bell:Deiters:Leal:2022]. 
+
 # Basic examples
 
 Loading the `aq17` thermodynamic dataset [@Miron2016;@Miron2017] and calculating standard properties of a substance (calcium aqueous ion $Ca^{2+}$). Output values are rounded for two decimal places.
