### 2022-12-22 3725f99
Added autodiff and Reaktoro bibentries
