+++ paper/paper.md
@@ -51,7 +51,7 @@ The ThermoFun library facilitates the computation of standard thermodynamic prop
 ![Example of standard thermodynamic properties of substances and reactions calculated using ThermoFun as a function of temperature along the saturated water vapor pressure curve (input parameters from `aq17` thermodynamic dataset [@Miron2016;@Miron2017]).\label{fig:properties}](figure1.png)
 
 The ThermoFun library can be linked to any C++ or Python geochemical equilibrium speciation, parameter optimization, or reactive mass-transport code as a source of standard thermodynamic properties of substances and/or reactions. As a standalone code, ThermoFun can be used simply for searching and tabulating thermodynamic properties of substances and reactions. For example, it has been used to calculate the standard properties of reactants and reactions at elevated temperatures when evaluating the performance of isocoulombic reactions, in order to extrapolate the properties of aqueous complexation reactions involving lanthanides and actinides to elevated temperatures [@Miron2020]. The library was also applied to evaluate the effect of using different equations of state for water-solvent on the calculated properties of dissolved aqueous species [@Miron2019]. 
-ThermoFun has already been adopted as a source of thermodynamic data in the [GEMS](https://gems.web.psi.ch/) equilibrium speciation codes [@Kulik2013] and the [Reaktoro](https://reaktoro.org) chemical modeling framework [@Leal2016].
+ThermoFun has already been adopted as a source of thermodynamic data in the [GEMS](https://gems.web.psi.ch/) equilibrium speciation codes [@Kulik2013] and the [Reaktoro](https://reaktoro.org) chemical modeling framework [@Leal:2015].
 
 # Features
 
@@ -65,7 +65,7 @@ The `ThermoBatch` class can be used to run batch calculations for a list of subs
 
 Reaction properties can be calculated for any `reaction` record present in the thermodynamic dataset, with the extrapolation to the given temperature and pressure done using the model code and its input parameters available in the record. Alternatively, the user can provide strings of reaction equations written using symbols of substances available in the thermodynamic dataset (e.g., Calcite $\leftrightharpoons$ Ca$^{2+}$ + CO$_{3}^{2-}$). In this case, the standard properties of reactions are calculated from the properties of reactants, each of which is first computed for the T, P of interest using the models and parameters defined in their respective `substance` records. Another option is to set a reaction with known properties to define the properties of a given substance. In this case, the standard properties of the reaction-defined substance at T, P are retrieved from the standard properties of the reaction calculated at T, P, along with the standard properties of the other reactants (substances) that are involved in the reaction. Thus, one thermodynamic dataset can contain a combination of substances, reaction-dependent substances, and reactions, whose consistent standard properties at T, P of interest are computed recursively.
 
-ThermoFun takes advantage of automatic differentiation for efficient and convenient calculation of derivatives. This feature allows ThermoFun to produce not only the standard thermodynamic properties of species and reactions, but also their temperature and pressure derivatives. This is necessary, for example, when chemical equilibrium problems where temperature and/or pressure are unknown, and Newton method is used to calculate these variables together with the amounts of species in equilibrium, as is done in Reaktoro [@ Leal2016]. Without temperature and pressure derivatives of standard Gibbs energies, the algorithm would not be fast and robust for these specific calculations. We plan to use the `autodiff` C++ library [@Leal:2018] in ThermoFun for even more capable autodiff capabilities, as it has been used in both Reaktoro [@Leal2016] and TEQP [@Bell:Deiters:Leal:2022]. 
[…]
