### 2022-12-24 27bdc1e
Fixes for JOSS manuscript (#57)
* fixes to manuscript text, added table with similar codes

* text fix

* Fixes author name

* Add bibentry for `Bell:Deiters:Leal:2022`

* Added TEQP as another code with purpose for thermo prop calculation

* Fixed alphabetical order (after adding TEQP)

* GEM-Selektor to codes with similar functionalities

Added GEM-Selektor to the list of codes with similar functionalities, fixed style in the paragraph above.

* Added paragraph under Features on automatic differentiation

* Added autodiff and Reaktoro bibentries

* Replaced Leal2016 by Leal:2015

Co-authored
[…]
