+++ paper/paper.bib
@@ -1,3 +1,30 @@
+@misc{Leal:2015,
+    author = {Leal, Allan M.M.},
+    howpublished = {\texttt{https://reaktoro.org}},
+    title = {{Reaktoro: An open-source unified framework for modeling chemically reactive systems}},
+    url = {https://reaktoro.org},
+    year = {2015}
+}
+@misc{Leal:2018,
+    author = {Leal, Allan M.M.},
+    howpublished = {\texttt{https://autodiff.github.io}},
+    title = {{{autodiff, a modern, fast and expressive {C++} library for automatic differentiatio}}n},
+    url = {https://autodiff.github.io},
+    year = {2018}
+}
+@article{Bell:Deiters:Leal:2022,
+    abstract = {This work uses advanced numerical techniques (complex differentiation and automatic differentiation) to efficiently and accurately compute all the required thermodynamic properties of an equation of state without any analytical derivatives─particularly without any handwritten derivatives. It avoids the tedious and error-prone process of symbolic differentiation, thus allowing for more rapid development of new thermodynamic models. The technique presented here was tested with several equations of state (van der Waals, Peng-Robinson, Soave-Redlich-Kwong, PC-SAFT, and cubic-plus-association) and high-accuracy multifluid models. A minimal set of algorithms (critical locus tracing and vapor-liquid equilibrium tracing) were implemented in an extensible and concise open-source C++ library: teqp (for Templated EQuation of state Package). This work demonstrates that highly complicated equations of state can be implemented faster yet with minimal computational overhead and negligible loss in numerical precision compared with the traditional approach that relies on analytical derivatives. We believe that the approach outlined in this work has the potential to establish a new computational standard when implementing computer codes for thermodynamic models.},
+    author = {Bell, Ian H. and Deiters, Ulrich K. and Leal, Allan M.M.},
+    doi = {10.1021/acs.iecr.2c00237},
+    file = {:C:/Users/Allan/Documents/Mendeley Desktop/Bell, Deiters, Leal - 2022 - Implementing an Equation of State without Derivatives teqp.pdf:BINARY},
+    issn = {15205045},
+    journal = {Industrial and Engineering Chemistry Research},
+    number = {17},
+    pages = {6010--6027},
+    title = {{Implementing an Equation of State without Derivatives: teqp}},
+    volume = {61},
+    year = {2022}
+}
 @article{Kulik2013,
[…]
