+++ docs/index.md
@@ -1,0 +1,68 @@
+<!--
+ Copyright (C) 2022 dmiron
+ 
+ This file is part of thermofun.
+ 
+ thermofun is free software: you can redistribute it and/or modify
+ it under the terms of the GNU General Public License as published by
+ the Free Software Foundation, either version 3 of the License, or
+ (at your option) any later version.
+ 
+ thermofun is distributed in the hope that it will be useful,
+ but WITHOUT ANY WARRANTY; without even the implied warranty of
+ MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ GNU General Public License for more details.
+ 
+ You should have received a copy of the GNU General Public License
+ along with thermofun.  If not, see <http://www.gnu.org/licenses/>.
+-->
+
+# [ThermoFun](https://thermohub.org/thermofun/thermofun/)
+
+#### Details on the API functionality can be found by [searching](https://docs.hdoc.io/dmiron/thermofun/search.html) the documentation. The naming is the  same in C++ and Python.
+
+Using ThermoFun is as simple as loading a [database](https://github.com/thermohub/thermofun-jupyter/blob/master/how-to-use-a-database.ipynb) initializing a [ThermoEngine](https://github.com/thermohub/thermofun-jupyter/blob/master/how-to-calculate-properties-thermoengine.ipynb)/[ThermoBatch](https://github.com/thermohub/thermofun-jupyter/blob/master/how-to-do-batch-calculations.ipynb) and calling the function to [calculate the properties](https://github.com/thermohub/thermofun-jupyter/blob/master/how-to-calculate-properties-substance-reaction.ipynb) for given substance(s)/reaction(s) at given T-P condition(s).
+
+```python
+  import thermofun as fun
+  database = fun.Database('aq17-thermofun.json')
+  engine = fun.ThermoEngine(database)
+  #                                        T(K)   P(Pa)  symbol
+  Ca_ion = engine.thermoPropertiesSubstance(473, 2000e5, 'Ca+2')
+  print(f'G0 {Ca_ion.gibbs_energy.val} J/mol')
+  #                                                     T(K)   P(Pa)  reaction equation
+  calcite_dissolution = engine.thermoPropertiesReaction(348.15, 1e5, 'Calcite = Ca+2 + CO3-2')
+  print(f'drS0 of (Cal = Ca+2 + CO3-2) is {calcite_dissolution.reaction_entropy.val}')
+  print(f'drG0 of (Cal = Ca+2 + CO3-2) is {calcite_dissolution.reaction_gibbs_energy.val}')
+  print(f'logK0 of (Cal = Ca+2 + CO3-2) is {calcite_dissolution.log_equilibrium_constant.val}')
+```
+
+In Python using the `help` function can provide hints on the usability.
+
+```python
+help(fun.ThermoEngine)
+
+Help on class ThermoEngine in module thermofun.PyThermoFun:
+
+class ThermoEngine(pybind11_builtins.pybind11_object)
+ |  Contains methods for calculating the thermodynamic properties of the substances and reactions
+ ...
+```
+
+Try the ThermoFun examples in your browser:
+
+[![Launch ThermoFun tutorials](https://img.shields.io/badge/launch-ThermoFun%20tutorials-brightgreen?style=for-the-badge&logo=jupyter)](https://mybinder.org/v2/gh/thermohub/thermofun-jupyter/master?urlpath=lab/)
+
+## The main ThermoFun classes
+
[…]
