### 2022-12-24 bdbc9ee
Imporved documentation (#58)
