+++ python/pyThermoFun/pyReaction.cpp
@@ -34,10 +34,10 @@ namespace ThermoFun {
 
 void exportReaction(py::module& m)
 {
-    py::class_<Reaction>(m, "Reaction")
+    py::class_<Reaction>(m, "Reaction", "A type used to describe a reaction")
         .def(py::init<>())
         .def(py::init<const Reaction&>())
-        .def(py::init<std::string>())
+        .def(py::init<std::string>(), "Constructor using a reaction record as a JSON string")
         .def("setName", &Reaction::setName,"Set the name of the Reaction.")
         .def("setSymbol", &Reaction::setSymbol,"Set the symbol of the Reaction.")
         .def("setEquation", &Reaction::setEquation,"Set the equation of the Reaction")
@@ -45,7 +45,7 @@ void exportReaction(py::module& m)
         .def("setReferenceT", &Reaction::setReferenceT,"Set reference temperature fo the standard properties of reaction (K)")
         .def("setReferenceP", &Reaction::setReferenceP,"Set references pressure for the standard properties of reaction (Pa)")
         .def("setLowerT", &Reaction::setLowerT,"Set upper temperature limit of the correction method (K)")
-        .def("setLowerP", &Reaction::setLowerP,"Set upper presure limit of the correction method (Pa)")
+        .def("setLowerP", &Reaction::setLowerP,"Set upper pressure limit of the correction method (Pa)")
         .def("setUpperT", &Reaction::setUpperT,"Set lower temperature limit of the correction method (K)")
         .def("setUpperP", &Reaction::setUpperP,"Set lower pressure limit for the correction method (Pa)")
         .def("setMethodGenEoS", &Reaction::setMethodGenEoS,"Set the code for general EOS calculation method")
@@ -64,7 +64,7 @@ void exportReaction(py::module& m)
         .def("referenceT", &Reaction::referenceT,"Returns the references temperature (K)")
         .def("referenceP", &Reaction::referenceP,"Returns the references pressure (Pa)")
         .def("lowerT", &Reaction::lowerT,"Return upper temperature limit of the correction method (K)")
-        .def("lowerP", &Reaction::lowerP,"Return upper presure limit of the correction method (Pa)")
+        .def("lowerP", &Reaction::lowerP,"Return upper pressure limit of the correction method (Pa)")
         .def("upperT", &Reaction::upperT,"Return lower temperature limit of the correction method (K)")
         .def("upperP", &Reaction::upperP,"Return lower pressure limit for the correction method (Pa)")
         .def("methodGenEOS", &Reaction::methodGenEOS,"Return the general EOS method code")
