### 2022-12-24 #4624 @gdmiron
@lucydot, I have managed to address the issues reported by @hgandhi2411 and @fnattino (thanks), as they follow:

* [[JOSS] Doc strings for python api thermohub/thermofun#52](https://github.com/thermohub/thermofun/issues/52)
The code documentation was updated. Docstrings are available in python with the use of `help` function, the generated code documentation is now available online and can be searched [here](https://docs.hdoc.io/dmiron/thermofun/).
* [[JOSS] - contributing guidelines thermohub/thermofun#53](https://github.com/thermohub/thermofun/issues/53)
These are now included in [readme](https://github.com/thermohub/thermofun/tree/docs#contributing-with-development)
* [[JOSS] - Python snippet in README thermohub/thermofun#54](https://github.com/thermohub/thermofun/issues/54)
Code snippets in readme were fixed to be consistent with latest code
* [[JOSS] - state of the field  thermohub/thermofun#55](https://github.com/thermohub/thermofun/issues/55)
A section `Codes with similar functionalities` was added to the manuscript

The manuscript as is now contains important general background for ThermoFun and on the importance and use of thermodynamic properties. It provides valuable context and helps to illustrate the relevance and importance of the library to a broad audience. Overall, the current length of the text is necessary to fully convey the features and capabilities of ThermoFun, as well as its potential value to researchers working in the fields of Earth and material science and engineering. As such, keeping the text at its current length will ensure that readers have a complete understanding of the library and its capabilities without them having to click away to pages on documentation and use examples.
