### 2023-01-06 #4624 @lucydot
Hi @gdmiron

The updates look good - I'll wait for the reviewers @hgandhi2411 @fnattino to give their verdicts (please tick off any boxes once you judge them complete, and let me know here when if/when you judge ready for acceptance @hgandhi2411 @fnattino).

Comments from an editorial viewpoint:
- `keeping the text at its current length will ensure that readers have a complete understanding of the library and its capabilities without them having to click away...` : JOSS papers are not meant to be a complete record, they are very much meant as adverts/pointers to the online documentation. Your paper should be closer to the advised length (250-1000 words). I know editing down words is a painful process, especially when it reads nicely as it is - however it is important that JOSS is consistent in this regard, and I cannot see a reason why this particular project requires a longer length text (after all most of the papers we publish relate to a research project with complex context/background)
- `Benefits of using ThermoFun` and `Features` have significant overlap. I suggest merging this into one smaller `Features` section - bullet points can help with conciseness.
- Personally I do like it when code usage examples are included in documentation, however if you cannot edit down to a more reasonable length, I suggest removing this section.
- `Statement of need` is important but at the moment it also contains quite a lot of background theory. I suggest that much of this could be included elsewhere (online) with a link in the paper text.
- You have included a comprehensive table of related projects - great. I suggest highlighting what ThermoFun offers differently from those listed - what is the "USP" of ThermoFun? This can be highlighted in the paper (if there is space after editing down) or elsewhere. You may want to also include the JOSS-published [`Surfinpy`](https://joss.theoj.org/papers/10.21105/joss.04014) in this list (I will it leave it to your judgement if there is enough overlap, may also be useful as an example of a very concise JOSS paper).
