### 2023-01-24 ef4d9a4
shortened manuscript to aorund 1000 words per request of JOSS editor
