+++ paper/paper.md
@@ -30,91 +30,31 @@ bibliography: paper.bib
 
 # Summary
 
-[ThermoFun](https://thermohub.org/thermofun) is an open source library that simplifies the calculation and retrieval of standard thermodynamic properties of chemical substances, compounds, and reactions among them at a range of temperatures and pressures. The library is developed in C++ for performance, but it also has a Python API for broader and more convenient usage. It employs a variety of thermodynamic models and equations of state for solid, aqueous, surface, gaseous, and molten substances, and their reactions with input properties and parameters from various thermodynamic datasets that are collected and curated in the in [ThermoHub](https://thermohub.org/thermohub) database. The library can be used as a standalone code for searching and tabulating thermodynamic properties or linked to other modeling codes that require standard thermodynamic data as input. It offers the flexibility to use different thermodynamic datasets, including custom datasets and datasets retrieved from the online ThermoHub database, and to choose the most suitable models for different classes of substances necessary in various modeling applications. It can serve as a common source of thermodynamic models for standard properties of substances and reactions that can be easily integrated and combined, significantly improving the modeling capabilities for diverse (geo)chemical systems and over wide ranges of conditions.
-
-# Benefits of using ThermoFun
-
-* It simplifies the calculation of standard thermodynamic properties of substances and reactions at a range of temperatures and pressures.
-* It includes a variety of equations of state and other thermodynamic models that can be used to calculate standard properties under different conditions, eliminating the need for the user to implement and test these models separately in their own code or spreadsheets.
-* It uses input properties and parameters from various thermodynamic datasets that are consistently kept in the online ThermoHub database, reducing the need to collect all necessary thermodynamic data from literature or to write complex scripts to import the data from different formats.
-* The thermodynamic models in ThermoFun are implemented and tested based on their original publications, and the thermodynamic data in ThermoHub is curated and documented by experts, saving time and effort and enabling more efficient use of geochemical code tools for scientific problems.
-* It serves as a common source of thermodynamic models for standard properties of substances and reactions that can be easily integrated and combined, significantly improving modeling capabilities for a wide range of (geo)chemical systems and conditions.
+[ThermoFun](https://thermohub.org/thermofun) is an open source library that facilitates the calculation and retrieval of standard thermodynamic properties of chemical substances, compounds, and reactions among them at a range of temperatures and pressures. The library is developed in C++ for performance, but it also has a Python API for broader and more convenient usage. It employs a variety of thermodynamic models and equations of state for solid, aqueous, surface, gaseous, and molten substances, and their reactions with input properties and parameters from various thermodynamic datasets that are collected and curated in the in [ThermoHub](https://thermohub.org/thermohub) database. The library can be used as a standalone code for searching and tabulating thermodynamic properties or linked to other modeling codes that require standard thermodynamic data as input. It offers the flexibility to use different thermodynamic datasets, including custom datasets and datasets retrieved from the online ThermoHub database, and to choose the most suitable models for different classes of substances necessary in various modeling applications. It can serve as a common source of thermodynamic models for standard properties of substances and reactions that can be easily integrated and combined, significantly improving the modeling capabilities for diverse (geo)chemical systems and over wide ranges of conditions.
 
 # Statement of need
 
[…]
