### 2023-01-24 #4624 @gdmiron
@lucydot Thanks for the comments and your time spent on this! The manuscript is now shortened to around 1100 words. The "USP" are better highlighted in the benefits section.
