### 2023-02-06 #4624 @lucydot
Ok @gdmiron, I'm pleased to say we are onto the final stages!

@hgandhi2411 @fnattino thank-you for your time and expertise reviewing for JOSS ✨

I've asked editorialbot to check the references. If you could confirm that DOIs are not available for those where not listed, and fix the url for the final reference (Zimmer et al) we can proceed to the next step.
