+++ paper/paper.bib
@@ -132,7 +132,7 @@ number = {7},
 pages = {899-947},
 year = {1992},
 issn = {0098-3004},
-doi = {https://doi.org/10.1016/0098-3004(92)90029-Q},
+doi = {10.1016/0098-3004(92)90029-Q},
 url = {https://www.sciencedirect.com/science/article/pii/009830049290029Q},
 author = {James W. Johnson and Eric H. Oelkers and Harold C. Helgeson},
 keywords = {SUPCRT92, Equations of state, Standard molal thermodynamic properties, Chemical equilibrium, Minerals, Gases, Aqueous species, HO, Thermodynamics, Geochemistry},
@@ -145,7 +145,7 @@ volume = {90},
 pages = {97-111},
 year = {2016},
 issn = {0098-3004},
-doi = {https://doi.org/10.1016/j.cageo.2016.02.013},
+doi = {10.1016/j.cageo.2016.02.013},
 url = {https://www.sciencedirect.com/science/article/pii/S0098300416300371},
 author = {Kurt Zimmer and Yilun Zhang and Peng Lu and Yanyan Chen and Guanru Zhang and Mehmet Dalkilic and Chen Zhu},
 keywords = {SUPCRT92, Arsenic, Geological carbon sequestration, Thermodynamic properties, Geochemical modeling, Chemical equilibrium},
@@ -158,7 +158,7 @@ volume = {606},
 pages = {120984},
 year = {2022},
 issn = {0009-2541},
-doi = {https://doi.org/10.1016/j.chemgeo.2022.120984},
+doi = {10.1016/j.chemgeo.2022.120984},
 url = {https://www.sciencedirect.com/science/article/pii/S0009254122002789},
 author = {Adedapo N. Awolayo and Benjamin M. Tutolo},
 keywords = {Helgeson-Kirkham-Flowers equation of state, LogK-density extrapolation, Solid-solutions, IAPWS95 equation of state, DEW model, Thermodynamic properties, Variable-chemistry clays, Geochemical modeling, Thermodynamic database},
@@ -240,7 +240,7 @@ pages = {597--643},
 volume = {89},
 number = {5},
 journal = {Pure and Applied Chemistry},
-doi = {doi:10.1515/pac-2016-1107},
+doi = {10.1515/pac-2016-1107},
 year = {2017},
 lastchecked = {2022-12-16}
 }
