### 2023-02-07 #4624 @gdmiron
Issues with DOI are fixed, thanks! I confirm that there are no DOI for references where not listed.
