### 2023-02-08 da2f852
GEMS-221: Fixed  build spdlog in header-only mode or as lib for thermofun project
