+++ cmake/modules/ThermoFunFindDeps.cmake
@@ -9,22 +9,22 @@ if(TFUN_BUILD_PYTHON)
     endif()
 endif()
 
-find_package(nlohmann_json REQUIRED)
-if(NOT nlohmann_json_FOUND)
-    message(FATAL_ERROR "nlohmann_json library not found")
-endif()
+#find_package(nlohmann_json REQUIRED)
+#if(NOT nlohmann_json_FOUND)
+#    message(FATAL_ERROR "nlohmann_json library not found")
+#endif()
 
 find_package(ChemicalFun REQUIRED)
 if(NOT ChemicalFun_FOUND)
     message(FATAL_ERROR "ChemicalFun library not found")
 endif()
 
-find_package(spdlog REQUIRED)
-if(NOT spdlog_FOUND)
-    message(FATAL_ERROR "spdlog not found")
+if(USE_SPDLOG_PRECOMPILED)
+   if(NOT TARGET spdlog::spdlog)
+       find_package(spdlog CONFIG REQUIRED)
+       if(NOT spdlog_FOUND)
+           message(FATAL_ERROR "spdlog not found")
+       endif()
+   endif()
 endif()
 
-find_package(fmt REQUIRED)
-if(NOT fmt_FOUND)
-    message(FATAL_ERROR "fmt not found")
-endif()
