### 2023-02-09 6fb3782
GEMS-221: Set spdlog in header-only mode for thermofun project, checked build&install
