+++ install-dependencies.sh
@@ -1,31 +1,30 @@
 #!/bin/bash
-# Installing dependencies needed to build ThermoFun on (k)ubuntu linux 16.04 or 18.04
+# Installing dependencies needed to build ThermoFun on (k)ubuntu linux 20.04
 
 EXTN=so
 
-#sudo rm -f /usr/local/include/nlohmann/json.hpp
-#sudo rm -rf /usr/local/include/eigen3/Eigen
-#sudo rm -rf /usr/local/include/pybind11
-#sudo rm -rf /usr/local/include/spdlog
-#sudo rm -rf /usr/local/include/fmt
-#sudo rm -f /usr/local/lib/libChemicalFun.$EXTN
+sudo rm -rf /usr/local/include/nlohmann
+sudo rm -rf /usr/local/include/eigen3/Eigen/Eigen
+sudo rm -rf /usr/local/include/pybind11
+sudo rm -rf /usr/local/include/spdlog
+sudo rm -f /usr/local/lib/libChemicalFun.$EXTN
 
 workfolder=${PWD}
-BRANCH_TFUN=master
+BRANCH_TFUN=spdlog-221
 BUILD_TYPE=Release
 
-# nlohmann/json
+# nlohmann/json library v. 3.6.1 or up
+# if not installed in /usr/local/include/nlohmann)
 test -f /usr/local/include/nlohmann/json.hpp || {
 
-	# Building json library
+	# Getting nlohmann/json library
 	mkdir -p ~/code && \
-                cd ~/code && \
+		cd ~/code && \
 		git clone https://github.com/nlohmann/json.git && \
 		cd json && \
 		mkdir -p build && \
 		cd build && \
-		cmake .. -DJSON_BuildTests=OFF && \
-		make && \
+		cmake .. -DCMAKE_BUILD_TYPE=Release -DJSON_BuildTests=OFF -DJSON_MultipleHeaders=ON && \
 		sudo make install
 
 	# Removing generated build files
@@ -33,66 +32,6 @@ test -f /usr/local/include/nlohmann/json.hpp || {
 		 rm -rf ~/code
 }
 
-# spdlog 
-# if no spdlog installed in /usr/local/lib/ (/usr/local/include/spdlog)
-test -d /usr/local/include/spdlog || {
-
-        # Building thermofun library
-        mkdir -p ~/code && \
-                cd ~/code && \
-                git clone https://github.com/gabime/spdlog -b v1.11.0  && \
-                cd spdlog && \
-                mkdir -p build && \
-                cd build && \
-                cmake .. -DCMAKE_BUILD_TYPE=Release -DSPDLOG_BUILD_TESTS=OFF -DCMAKE_INSTALL_LIBDIR=lib -DSPDLOG_BUILD_SHARED=ON -DSPDLOG_FMT_EXTERNAL=ON \
-                make && \
-                sudo make install
-
-        # Removing generated build files
-        cd ~ && \
-                 rm -rf ~/code
-}
-
-# fmt 
-# if no spdlog installed in /usr/local/lib/ (/usr/local/include/fmt)
-test -d /usr/local/include/fmt || {
-
-        # Building thermofun library
-        mkdir -p ~/code && \
-                cd ~/code && \
-                git clone hhttps://github.com/fmtlib/fmt.git -b v9.1.0  && \
-                cd fmt && \
-                mkdir -p build && \
-                cd build && \
-                cmake .. \
-                make && \
-                sudo make install
-
-        # Removing generated build files
-        cd ~ && \
-                 rm -rf ~/code
-}
-
-#Pybind11
-test -d /usr/local/include/pybind11 || {
-
-        # Building pybind11 library
-        mkdir -p ~/code && \
-                cd ~/code && \
-                git clone https://github.com/pybind/pybind11.git && \
[…]
