### 2023-02-16 #4624 @lucydot
Paper is looking good! If you could do the following @gdmiron:
- please adjust the zenodo repo metadata so that the title is an exact match to the title of your paper: "ThermoFun: A C++/Python library for computing standard thermodynamic properties of substances and reactions across wide ranges of temperatures and pressures"
- We also ask that the author list is an exact match to the author list on the paper. Currently there are an additional two authors on the zenodo repo.
- please confirm version number (v0.3.9) is correct
