### 2023-02-20 #4624 @xuanxu
The error happens because there are two different bib entries qith the same name (Leal2017): [here](https://github.com/thermohub/thermofun/blob/paper/paper/paper.bib#L72) and [here](https://github.com/thermohub/thermofun/blob/paper/paper/paper.bib#L234)
