### 2023-02-21 d10ee97
removed duplicate bib entry
