+++ paper/paper.bib
@@ -231,19 +231,6 @@ year = {2021},
 title={Thermo: Chemical properties component of Chemical Engineering Design Library (ChEDL)},
 url = {https://github.com/CalebBell/thermo}
 }
-@article{Leal2017,
-url = {https://doi.org/10.1515/pac-2016-1107},
-title = {An overview of computational methods for chemical equilibrium and kinetic calculations for geochemical and reactive transport modeling},
-title = {},
-author = {Allan M. M. Leal and Dmitrii A. Kulik and William R. Smith and Martin O. Saar},
-pages = {597--643},
-volume = {89},
-number = {5},
-journal = {Pure and Applied Chemistry},
-doi = {10.1515/pac-2016-1107},
-year = {2017},
-lastchecked = {2022-12-16}
-}
 @software{cantera,
     author = "David G. Goodwin and Harry K. Moffat and Ingmar Schoegl and Raymond L.
               Speth and Bryan W. Weber",
