### 2023-02-21 #4624 @gdmiron
@lucydot I removed the duplicate bib entry.
