### 2023-02-28 #4624 @kyleniemeyer
Hi @gdmiron, just doing final checks on the paper before publishing.

One quibble in the Statement of Need:
> It requires the standard thermodynamic properties of all involved substances (species) or reactions at a specific temperature (T) and pressure (P) as input.

This isn't *technically* true in general, since the thermodynamic modeling can also use equations of state that relate intensive properties. Could you edit to be more generally correct, versus what specifically applies to your software package?
