### 2023-03-01 #4624 @gdmiron
@kyleniemeyer thanks for making the final check.

This part is related to the previous statement

> Thermodynamic modeling is a powerful tool that allows **for the investigation of equilibrium speciation in chemical systems** under conditions and time scales that cannot be reached in laboratory settings.

Then the following text refers to "the investigation of equilibrium speciation in chemical systems" for which these properties are required.

To make it more clear, should I change the text to the example below, or just change "It requires" with "This requires"

> Thermodynamic modeling is a powerful tool that allows for the investigation of equilibrium speciation in chemical systems under conditions and time scales that cannot be reached in laboratory settings. For this type of calculation, it requires the standard thermodynamic properties of all involved substances (species) or reactions at a specific temperature (T) and pressure (P) as input.
