### 2023-03-01 #4624 @kyleniemeyer
@gdmiron ah, that makes more sense. How about just "This type of calculation requires" for the second sentence?
