### 2023-03-01 be78d92
clarify application to equilibrium speciation
