+++ paper/paper.md
@@ -34,7 +34,7 @@ bibliography: paper.bib
 
 # Statement of need
 
-Thermodynamic modeling is a powerful tool that allows for the investigation of equilibrium speciation in chemical systems under conditions and time scales that cannot be reached in laboratory settings. It requires the standard thermodynamic properties of all involved substances (species) or reactions at a specific temperature (T) and pressure (P) as input. Equations of state are used to compute these properties, known as standard thermodynamic models. However, existing thermodynamic modeling codes and standalone codes may not be directly compatible, making it difficult to retrieve properties calculated with different methods and use them in different codes, or to compare and tabulate thermodynamic data from different sources. The ThermoFun library simplifies this process by facilitating the computation of standard thermodynamic properties of substances and reactions (see \autoref{fig:properties}), and providing fast and simple access to many thermodynamic datasets, methods, and equations of state. This allows for more efficient use of these tools in modeling codes, or for the optimization and fitting of model parameters, evaluation of the models' performance, range of applicability, and accuracy. Additionally, ThermoFun is extendable with new models and has remote access to a curated collection of thermodynamic datasets on ThermoHub, making it a useful tool for various studies on hydrothermal processes, cementitious materials, nuclear engineering, waste incineration, radioactive and chemical management and disposal, and other (geo)chemical environments.
+Thermodynamic modeling is a powerful tool that allows for the investigation of equilibrium speciation in chemical systems under conditions and time scales that cannot be reached in laboratory settings. This type of calculation requires the standard thermodynamic properties of all involved substances (species) or reactions at a specific temperature (T) and pressure (P) as input. Equations of state are used to compute these properties, known as standard thermodynamic models. However, existing thermodynamic modeling codes and standalone codes may not be directly compatible, making it difficult to retrieve properties calculated with different methods and use them in different codes, or to compare and tabulate thermodynamic data from different sources. The ThermoFun library simplifies this process by facilitating the computation of standard thermodynamic properties of substances and reactions (see \autoref{fig:properties}), and providing fast and simple access to many thermodynamic datasets, methods, and equations of state. This allows for more efficient use of these tools in modeling codes, or for the optimization and fitting of model parameters, evaluation of the models' performance, range of applicability, and accuracy. Additionally, ThermoFun is extendable with new models and has remote access to a curated collection of thermodynamic datasets on ThermoHub, making it a useful tool for various studies on hydrothermal processes, cementitious materials, nuclear engineering, waste incineration, radioactive and chemical management and disposal, and other (geo)chemical environments.
 
 ![Example of standard thermodynamic properties of substances and reactions calculated using ThermoFun as a function of temperature along the saturated water vapor pressure curve (input parameters from `aq17` thermodynamic dataset [@Miron2016;@Miron2017]).\label{fig:properties}](figure1.png)
 
